organize tests to match new layout
