--- src/test/java/com/marginallyclever/robotoverlord/EntityTest.java
@@ -1,22 +1,0 @@
-package com.marginallyclever.robotoverlord;
-
-import com.marginallyclever.robotoverlord.entity.Entity;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
-
-public class EntityTest {
-    private static void saveAndLoad(Entity a, Entity b) throws Exception {
-        b.parseJSON(a.toJSON());
-        Assertions.assertEquals(a.toString(),b.toString());
-    }
-
-    @Test
-    public void saveAndLoad() throws Exception {
-        Entity a = new Entity();
-        Entity b = new Entity();
-        EntityTest.saveAndLoad(a,b);
-
-        a.setExpanded(!a.getExpanded());
-        EntityTest.saveAndLoad(a,b);
-    }
-}
