+++ src/test/java/com/marginallyclever/robotoverlord/components/ComponentFactoryTest.java
@@ -1,7 +1,5 @@
-package com.marginallyclever.robotoverlord;
+package com.marginallyclever.robotoverlord.components;
 
-import com.marginallyclever.robotoverlord.components.Component;
-import com.marginallyclever.robotoverlord.components.ComponentFactory;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
