+++ src/test/java/com/marginallyclever/robotoverlord/components/ComponentTest.java
@@ -1,6 +1,5 @@
-package com.marginallyclever.robotoverlord;
+package com.marginallyclever.robotoverlord.components;
 
-import com.marginallyclever.robotoverlord.components.Component;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
