+++ src/test/java/com/marginallyclever/robotoverlord/components/shapes/MeshFromFileTest.java
@@ -1,13 +1,10 @@
 package com.marginallyclever.robotoverlord.components.shapes;
 
-import com.marginallyclever.robotoverlord.ComponentTest;
+import com.marginallyclever.robotoverlord.components.ComponentTest;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 import java.io.File;
-import java.net.URL;
-import java.nio.file.Path;
-import java.nio.file.Paths;
 
 public class MeshFromFileTest {
     @Test
