+++ src/test/java/com/marginallyclever/robotoverlord/systems/render/ViewportTest.java
@@ -1,4 +1,4 @@
-package com.marginallyclever.robotoverlord;
+package com.marginallyclever.robotoverlord.systems.render;
 
 import com.marginallyclever.robotoverlord.systems.render.Viewport;
 import org.junit.jupiter.api.Assertions;
